+++ lib/Service/python/tests/dev_test.py
@@ -4,7 +4,11 @@ Currently it is a dev file for some features testing. Not included in releases, 
 
 import sys
 import platform
-from lib.Service.python import install
+import os
+from pathlib import Path
+mdc_py_path = str(Path(os.path.dirname(os.path.abspath(__file__))).parent)
+sys.path.append(mdc_py_path)
+import install
 
 
 # @copyright Copyright (c) 2021 Andrey Borysenko <andrey18106x@gmail.com>