+++ src/mixins/Configure.js
@@ -24,7 +24,7 @@
 
 import { generateUrl } from '@nextcloud/router'
 import axios from '@nextcloud/axios'
-import { showError, showSuccess } from '@nextcloud/dialogs'
+import { showError, showSuccess, showWarning } from '@nextcloud/dialogs'
 
 export default {
 	created() {
@@ -39,15 +39,15 @@ export default {
 		return {
 			installed_setting: null,
 			installed: false,
+			installed_list: [],
+			not_installed_list: {},
+			video_required: [],
+			available_algorithms: [],
+			errors: [],
+			warnings: [],
 			installing: false,
 			checking: false,
 			updating: false,
-			installed_packages_list: {},
-			available_algorithms: [],
-			video_required: [],
-			packages_list: [],
-			errors: [],
-			warnings: [],
 		}
 	},
 	methods: {
@@ -64,11 +64,11 @@ export default {
 					this.installed_setting = res.data.setting
 					this.installed_setting.value = JSON.parse(this.installed_setting.value)
 					this.installed = res.data.setting.value.status
-					this.packages_list = this.installed_setting.value.list
-					this.installed_packages_list = {
-						required: this.installed_setting.value.installed_list.required,
-						optional: this.installed_setting.value.installed_list.optional,
-						boost: this.installed_setting.value.installed_list.boost,
+					this.installed_list = this.installed_setting.value.installed_list
+					this.not_installed_list = {
+						required: this.installed_setting.value.not_installed_list.required,
+						optional: this.installed_setting.value.not_installed_list.optional,
+						boost: this.installed_setting.value.not_installed_list.boost,
 					}
 					this.available_algorithms = this.installed_setting.value.available_algorithms
 					this.video_required = this.installed_setting.value.video_required
@@ -83,19 +83,20 @@ export default {
 		async install() {
 			this.installing = true
 			axios.get(generateUrl('/apps/mediadc/api/v1/python/install')).then(res => {
-				if (res.data.success) {
-					this.installed = true
-					this.parsePythonResponseData(res)
-					this.updateInstalledSetting().then(() => {
-						this.installing = false
-						showSuccess(t('mediadc', 'Installation successfully finished!'))
-					}).catch(() => {
-						this.installing = false
-						showError(t('mediadc', 'Installation failed. Try again.'))
-					})
-				} else {
+				this.parsePythonResponseData(res)
+				this.updateInstalledSetting().then(() => {
 					this.installing = false
-				}
+					if (res.data.success) {
+						showSuccess(t('mediadc', 'Installation successfully finished'))
+					} else if (!res.data.success && res.data.installed) {
+						showWarning(t('mediadc', 'Installation finished. Not all packages installed'))
+					} else {
+						showError(t('mediadc', 'Installation failed. Try again.'))
+					}
+				}).catch(() => {
+					this.installing = false
+					showError(t('mediadc', 'Installation failed. Try again.'))
+				})
 			})
 		},
 		async updateInstalledSetting() {
@@ -103,11 +104,15 @@ export default {
 		},
 		async installDepsList(listName) {
 			this.updating = true
-			axios.post(generateUrl(`/apps/mediadc/api/v1/python/install/${listName}`)).then(res => {
+			axios.post(generateUrl('/apps/mediadc/api/v1/python/install'), { listName }).then(res => {
 				this.parsePythonResponseData(res)
 				this.updateInstalledSetting().then(() => {
 					this.updating = false
-					showSuccess(t('mediadc', 'Package list successfully installed'))
+					if (res.data.success) {
+						showSuccess(t('mediadc', 'Package list successfully installed'))
+					} else {
+						showError(t('mediadc', 'Package list installation failed'))
+					}
 				}).catch(() => {
 					this.updating = false
 					showError(t('mediadc', 'Package list installation failed'))
@@ -117,20 +122,35 @@ export default {
 		async deleteDepsList(listName) {
 			this.updating = true
 			axios.post(generateUrl('/apps/mediadc/api/v1/python/delete'),
-				{ packagesList: Object.values(this.packages_list[listName]).filter(item => item.location !== 'global').map(item => item.package) }).then(res => {
-				this.updating = false
+				{ packagesList: Object.values(this.installed_list[listName]).filter(item => item.location !== 'global').map(item => item.package) }).then(res => {
 				this.parsePythonResponseData(res)
-				this.updateInstalledSetting()
+				this.updateInstalledSetting().then(() => {
+					this.updating = false
+					if (res.data.success) {
+						showSuccess(t('mediadc', 'Package list successfully deleted'))
+					} else {
+						showError(t('mediadc', 'Some error occured while deleting package list'))
+					}
+				})
+			}).catch(err => {
+				console.debug(err)
+				this.updating = false
+				showError(t('mediadc', 'Some error occured while deleting packages'))
 			})
 		},
 		async updateDepsList(listName) {
 			this.updating = true
 			axios.post(generateUrl('/apps/mediadc/api/v1/python/update'),
-				{ packagesList: Object.values(this.packages_list[listName]).filter(item => item.location !== 'global').map(item => item.package) }).then(res => {
-				this.updating = false
+				{ packagesList: Object.values(this.installed_list[listName]).filter(item => item.location !== 'global').map(item => item.package) }).then(res => {
 				this.parsePythonResponseData(res)
-				this.updateInstalledSetting()
-				showSuccess(t('mediadc', 'Packages successfully updated'))
+				this.updateInstalledSetting().then(() => {
+					this.updating = false
+					if (res.data.success) {
+						showSuccess(t('mediadc', 'Packages successfully updated'))
+					} else {
+						showError(t('mediadc', 'Packages update failed. Try again.'))
+					}
+				})
 			}).catch(err => {
 				console.debug(err)
 				showError(t('mediadc', 'Packages update failed'))
@@ -139,9 +159,19 @@ export default {
 		async check() {
 			this.checking = true
 			axios.get(generateUrl('/apps/mediadc/api/v1/python/check')).then(res => {
-				this.checking = false
 				this.parsePythonResponseData(res)
-				this.updateInstalledSetting()
+				this.updateInstalledSetting().then(() => {
+					this.checking = false
+					if (res.data.installed) {
+						showSuccess(t('mediadc', 'All required dependencies installed'))
+					} else {
+						showError(t('mediadc', 'Not all required packages installed'))
+					}
+				}).catch(err => {
+					this.checking = false
+					console.debug(err)
+					showError(t('mediadc', 'Dependencies checking failed. Try again.'))
+				})
 			}).catch(err => {
 				console.debug(err)
 				showError(t('mediadc', 'Dependencies checking failed. Try again.'))
@@ -151,8 +181,8 @@ export default {
 		parsePythonResponseData(res) {
 			this.installed = res.data.installed
 			this.installed_setting.value.status = res.data.installed
-			this.installed_setting.value.list = res.data.list
-			this.installed_setting.value.installed_list = {
+			this.installed_setting.value.installed_list = res.data.installed_list
+			this.installed_setting.value.not_installed_list = {
 				required: res.data.required,
 				optional: res.data.optional,
 				boost: res.data.boost,
@@ -160,8 +190,8 @@ export default {
 			this.installed_setting.value.available_algorithms = res.data.available_algorithms
 			this.installed_setting.value.video_required = res.data.video_required
 			this.$store.dispatch('setSetting', this.installed_setting)
-			this.packages_list = res.data.list
-			this.installed_packages_list = {
+			this.installed_list = res.data.installed_list
+			this.not_installed_list = {
 				required: res.data.required,
 				optional: res.data.optional,
 				boost: res.data.boost,