+++ CHANGELOG.md
@@ -2,16 +2,46 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.0 - yyyy-mm-dd]
-
-### Added
-
-- What added
-
-### Changed
-
-- What changed
+## [0.1.4 - 2021-10-07]
 
 ### Fixed
 
-- What fixed
+- Proper work with non ascii characters in files names. Thanks to kovge.
+
+## [0.1.3 - 2021-10-04]
+
+### Fixed
+
+- Now properly parse configs for database connect.
+
+## [0.1.2 - 2021-09-30]
+
+### Added
+
+- Added image and video previews support
+
+### Fixed
+
+- Fixed not disabled buttons for impossible actions on Configuration page (update/delete on not installed packages)
+- Now properly display installation errors
+
+## [0.1.1 - 2021-09-26]
+
+### Added
+
+### Changed
+
+### Fixed
+
+- Fixed opening photo&video via Viewer in Nextcloud 21, 22
+
+## [0.1.0 - 2021-09-23]
+
+### Added
+
+### Changed
+
+- First beta release available at Nextcloud App Store
+
+### Fixed
+